requirements.txt
flask
gunicorn
discord.py